requirements.txt
fastapi
uvicorn
pydantic
pandas
snowflake-connector-python
